Add support for -beta-x suffix when parsing hazelcast version (#12365)
Adds support for -beta-x suffix, e.g. 3.10-beta-1.

(cherry picked from commit 62be29a)
