+++ hazelcast/src/main/java/com/hazelcast/util/StringUtil.java
@@ -56,8 +56,8 @@ public final class StringUtil {
     /**
      * Pattern used to tokenize version strings.
      */
-    private static final Pattern VERSION_PATTERN
-            = Pattern.compile("^([\\d]+)\\.([\\d]+)(\\.([\\d]+))?(-[\\w]+)?(-SNAPSHOT)?$");
+    public static final Pattern VERSION_PATTERN
+            = Pattern.compile("^(\\d+)\\.(\\d+)(\\.(\\d+))?(-\\w+(?:-\\d+)?)?(-SNAPSHOT)?$");
 
     private static final String GETTER_PREFIX = "get";
