+++ hazelcast/src/test/java/com/hazelcast/version/MemberVersionTest.java
@@ -38,6 +38,8 @@ public class MemberVersionTest {
 
     private static final String VERSION_3_8_SNAPSHOT_STRING = "3.8-SNAPSHOT";
     private static final String VERSION_3_8_1_RC1_STRING = "3.8.1-RC1";
+    private static final String VERSION_3_8_1_BETA_1_STRING = "3.8.1-beta-1";
+    private static final String VERSION_3_8_BETA_2_STRING = "3.8-beta-2";
     private static final String VERSION_3_8_2_STRING = "3.8.2";
     private static final String VERSION_3_9_0_STRING = "3.9.0";
     private static final String VERSION_UNKNOWN_STRING = "0.0.0";
@@ -58,6 +60,8 @@ public class MemberVersionTest {
         assertTrue(MemberVersion.UNKNOWN.isUnknown());
         assertFalse(MemberVersion.of(VERSION_3_8_SNAPSHOT_STRING).isUnknown());
         assertFalse(MemberVersion.of(VERSION_3_8_1_RC1_STRING).isUnknown());
+        assertFalse(MemberVersion.of(VERSION_3_8_1_BETA_1_STRING).isUnknown());
+        assertFalse(MemberVersion.of(VERSION_3_8_BETA_2_STRING).isUnknown());
         assertFalse(MemberVersion.of(VERSION_3_8_2_STRING).isUnknown());
     }
 
@@ -70,6 +74,20 @@ public class MemberVersionTest {
     public void testVersionOf_whenVersionStringIsSnapshot() {
         MemberVersion expected = MemberVersion.of(3, 8, 0);
         assertEquals(expected, MemberVersion.of(VERSION_3_8_SNAPSHOT_STRING));
+    }
+
+    @Test
+    public void testVersionOf_whenVersionStringIsBeta() {
+        assertEquals(MemberVersion.of(3, 8, 0), MemberVersion.of(VERSION_3_8_BETA_2_STRING));
+        assertEquals(MemberVersion.of(3, 8, 1), MemberVersion.of(VERSION_3_8_1_BETA_1_STRING));
+    }
+
+    @Test
+    public void test_constituents_whenVersionStringIsBeta() {
+        final MemberVersion expected = MemberVersion.of(VERSION_3_8_BETA_2_STRING);
+        assertEquals(3, expected.getMajor());
+        assertEquals(8, expected.getMinor());
+        assertEquals(0, expected.getPatch());
     }
 
     @Test
